added universal form
