+++ src/components/Feeds/HotFeed.js
@@ -1,5 +1,5 @@
 import { useCollection } from 'react-firebase-hooks/firestore';
-import {collection, orderBy, query, serverTimestampmp} from 'firebase/firestore';
+import {collection, orderBy, query} from 'firebase/firestore';
 import Error from '../shared/Error.js'
 import Post from '../../pages/Post';
