+++ src/pages/Submit/PostStart.js
@@ -1,4 +1,4 @@
-import Form from "./Form";    
+import Form from "../../components/submit/functions/Form";    
 export default class PostStrat extends Form{
     constructor(formValue, user, messageRef, selectedGenre, file){
         super(formValue, user, messageRef,selectedGenre,file)
@@ -14,17 +14,3 @@ export default class PostStrat extends Form{
     }
 } 
 
-export class QandaStrat extends Form{
-    constructor(formValue, user, messageRef, author, file){
-        super(formValue, user, messageRef,author,file)
-        this.post = {
-            author: author,
-        }
-       
-    }
-    async sendMessage(payload){
-  
-        this.payload = Object.assign({}, this.payload, this.post)
-        return await super.sendMessage(this.payload)
-    }
-}
